+++ src/Components/Forms/Components/index.ts
@@ -1,1 +1,2 @@
 export { default as FormBuilder } from './FormBuilder';
+export { default as TextInput } from './FormBuilder';