+++ src/Pages/Code/Code.tsx
@@ -3,10 +3,16 @@ import { Wrapper, Header, Section, MultiTable } from 'Components';
 import { useContracts } from 'redux/hooks';
 // @ts-ignore
 import { useParams } from 'react-router-dom';
-import { CodeDetails, CodeContracts } from './Components';
+import { CodeDetails, CodeContracts, CodeTxs } from './Components';
 
 const Code = () => {
-  const { getContractCode, getContractsByCode, contractsByCodeTotal } = useContracts();
+  const { 
+    getContractCode, 
+    getContractsByCode, 
+    contractsByCodeTotal, 
+    codeTxsTotal,
+    getCodeTxs,
+  } = useContracts();
   const { codeId } = useParams();
 
   const [activeTableTab, setActiveTableTab] = useState(0);
@@ -14,7 +20,13 @@ const Code = () => {
   useEffect(() => {
     getContractCode({ id: codeId });
     getContractsByCode({ id: codeId, page: 1, count: 30 });
-  }, [codeId, getContractCode, getContractsByCode]);
+    getCodeTxs({
+      id: codeId,
+      page: 1,
+      count: 30,
+      status: '',
+    });
+  }, [codeId, getContractCode, getContractsByCode, getCodeTxs]);
 
   return (
     <Wrapper>
@@ -26,6 +38,7 @@ const Code = () => {
         <MultiTable active={activeTableTab} setActive={setActiveTableTab}>
           <CodeDetails key="Details"/>
           <CodeContracts key={`Contracts (${contractsByCodeTotal})`} />
+          <CodeTxs key={`Transactions (${codeTxsTotal})`} />
         </MultiTable>
       </Section>
     </Wrapper>