+++ src/theme/Colors/Colors.js
@@ -105,6 +105,7 @@ export const Colors = {
   TOGGLE_NOTCH: allColors.WHITE,
   // Warnings
   WARNING_BORDER: allColors.RED_PRIMARY,
+  WARNING_BORDER_LIGHTER: allColors.ORANGE_DARK,
   // +/- indicators
   POSITIVE_CHANGE: allColors.GREEN_POSITIVE_PRIMARY,
   NEGATIVE_CHANGE: allColors.RED_NEGATIVE_PRIMARY,