+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix filter for transactions chart #313
 - Fix proposal denom to use nhash #320
 - Fix some FE formatting issues #319
+- Fix some smart contracts issues seen in test #325
 
 ### Features
 - Update asset price per unit calcs for new data object #296